schedule page added + file naming changes
