+++ src/vorta/assets/UI/logpage.ui
