+++ src/vorta/assets/UI/shellcommandspage.ui
