+++ src/vorta/views/log_page.py
@@ -9,7 +9,7 @@ from vorta import config
 from vorta.store.models import EventLogModel
 from vorta.utils import get_asset
 
-uifile = get_asset('UI/logtablewidget.ui')
+uifile = get_asset('UI/logpage.ui')
 LogTableUI, LogTableBase = uic.loadUiType(uifile)
 
 
@@ -21,7 +21,7 @@ class LogTableColumn:
     ReturnCode = 4
 
 
-class LogTableWidget(LogTableBase, LogTableUI):
+class LogPanel(LogTableBase, LogTableUI):
     def __init__(self, parent=None):
         super().__init__(parent)
         self.setupUi(self)
