+++ src/vorta/views/networks_page.py
@@ -5,11 +5,11 @@ from PyQt6.QtWidgets import QCheckBox, QLabel, QListWidget, QListWidgetItem
 from vorta.store.models import BackupProfileMixin, WifiSettingModel
 from vorta.utils import get_asset, get_sorted_wifis
 
-uifile = get_asset('UI/networkspanel.ui')
+uifile = get_asset('UI/networkspage.ui')
 NetworksUI, NetworksBase = uic.loadUiType(uifile)
 
 
-class NetworksPanel(NetworksBase, NetworksUI, BackupProfileMixin):
+class NetworksPage(NetworksBase, NetworksUI, BackupProfileMixin):
     def __init__(self, parent=None):
         super().__init__(parent)
         self.setupUi(self)
@@ -26,9 +26,11 @@ class NetworksPanel(NetworksBase, NetworksUI, BackupProfileMixin):
         self.wifiListWidget.itemChanged.connect(self.save_wifi_item)
 
     def on_metered_networks_state_changed(self, state):
+        profile = self.profile()
         attr = 'dont_run_on_metered_networks'
         new_value = state != Qt.CheckState.Checked
         self.save_profile_attr(attr, new_value)
+        self.meteredNetworksCheckBox.setChecked(False if profile.dont_run_on_metered_networks else True)
 
     def populate_wifi(self):
         self.wifiListWidget.clear()
